Extracted common test cases to a separate class
